feat: 完成牌组模块的大部分功能
牌组模块开发告一段落，下一步就是复习模块，复习模块开发完了之后就是弄前端了

+++ api/v1/card.go
@@ -23,5 +23,5 @@ type CardRequest struct {
 }
 
 type CardIdReq struct {
-	Ids []int `json:"ids"`
+	Ids []int64 `json:"ids"`
 }
